+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+main:
+  * Add `--usb-mode` option to `aravis-capture`.
+  * Expose features from `aravis-sys` in `aravis`.
+
 v0.8.2 - 2023-01-02:
   * Add support for Aravis `v0.8.22`.